requirements.txt
Pillow==9.1.1
requests==2.27.1
reportlab~=3.6.9